requirements.txt
astroid==2.8.0
backports.entry-points-selectable==1.1.0
black==21.9b0
cfgv==3.3.1
click==8.0.1
distlib==0.3.3
filelock==3.3.0
gevent==21.8.0
greenlet==1.1.2
identify==2.3.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.4.0
pre-commit==2.15.0
pydantic==1.8.2
pylint==2.11.1
PyYAML==5.4.1
pyzmq==22.3.0
regex==2021.9.30
six==1.16.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
virtualenv==20.8.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
